Factor the country detection code out into it's own module
